+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/graph/BaseBinding.kt
@@ -20,14 +20,17 @@ internal interface BaseBinding<
   val isTransient: Boolean
     get() = false
 
+  /**
+   * Some types may be implicitly deferrable such as lazy/provider types, instance-based bindings,
+   * or bindings that don't participate in object construction such as object classes or members
+   * injectors.
+   */
   val isImplicitlyDeferrable: Boolean
     get() = contextualTypeKey.isDeferrable
 
   fun renderLocationDiagnostic(short: Boolean = false): LocationDiagnostic
+
   fun renderDescriptionDiagnostic(short: Boolean = false, underlineTypeKey: Boolean = false): String
 }
 
-internal data class LocationDiagnostic(
-  val location: String,
-  val description: String?,
-)
+internal data class LocationDiagnostic(val location: String, val description: String?)