+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/ProviderFieldCollector.kt
@@ -15,16 +15,29 @@ internal class ProviderFieldCollector(private val graph: IrBindingGraph) {
     val needsField: Boolean
       get() {
         // Scoped, graph, and members injector bindings always need provider fields
-        if (binding.isScoped()) return true
-        if (binding is IrBinding.GraphDependency) return true
-        if (binding is IrBinding.MembersInjected && !binding.isFromInjectorFunction) return true
-        // Multibindings are always created adhoc
-        if (binding is IrBinding.Multibinding) return false
-        // Assisted types always need to be a single field to ensure use of the same provider
-        if (binding is IrBinding.Assisted) return true
-        // TODO what about assisted but no assisted params? These also don't become providers
-        //  we would need to track a set of assisted targets somewhere
-        if (binding is IrBinding.ConstructorInjected && binding.isAssisted) return true
+        if (binding.isScoped()) {
+          return true
+        }
+
+        when (binding) {
+          is IrBinding.GraphDependency,
+          // Assisted types always need to be a single field to ensure use of the same provider
+          is IrBinding.Assisted -> {
+            return true
+          }
+          // TODO what about assisted but no assisted params? These also don't become providers
+          //  we would need to track a set of assisted targets somewhere
+          is IrBinding.ConstructorInjected -> {
+            return binding.isAssisted
+          }
+          // Multibindings are always created adhoc
+          is IrBinding.Multibinding -> {
+            return false
+          }
+          else -> {
+            // Do nothing
+          }
+        }
 
         if (
           binding.typeKey.qualifier?.ir?.annotationClass?.classId ==